requirements.txt
black==21.6b0
Django==3.*
django-environ==0.4.5
djangorestframework==3.*
freezegun==1.1.0
factory_boy==3.2.0
pytest==6.2.4
pytest-django==4.4.0
stripe==2.*